Introduce admin bar menu item for admins
The admin bar menu item has an icon that serves as a notification for the
admin that the current site's protection is active.

This implementation resulted in the new `guard_get_protection_details()`
which replaces the previous `guard_network_sites_protect_details` filter
that was used in the Sites network admin page.